Move passToFoo() functions to their own file
